fix(semantic): panics when function is the part of `IfStatement` (#12190)
close: #10015

Also, removed weird SymbolId!
