+++ crates/oxc_semantic/src/binder.rs
@@ -1,18 +1,13 @@
 //! Declare symbol for `BindingIdentifier`s
 
-use std::ptr;
-
 use oxc_allocator::{Address, GetAddress};
 use oxc_ast::{AstKind, ast::*};
 use oxc_ecmascript::{BoundNames, IsSimpleParameterList};
 use oxc_span::GetSpan;
-use oxc_syntax::{
-    node::NodeId,
-    scope::{ScopeFlags, ScopeId},
-    symbol::SymbolFlags,
-};
+use oxc_syntax::{node::NodeId, scope::ScopeFlags, symbol::SymbolFlags};
 
 use crate::SemanticBuilder;
+use crate::checker::is_function_part_of_if_statement;
 
 pub trait Binder<'a> {
     #[expect(unused_variables)]
@@ -138,27 +133,6 @@ impl<'a> Binder<'a> for Class<'a> {
     }
 }
 
-/// Check for Annex B `if (foo) function a() {} else function b() {}`
-fn is_function_part_of_if_statement(function: &Function, builder: &SemanticBuilder) -> bool {
-    if builder.current_scope_flags().is_strict_mode() {
-        return false;
-    }
-    let AstKind::IfStatement(stmt) = builder.nodes.parent_kind(builder.current_node_id) else {
-        return false;
-    };
-    if let Statement::FunctionDeclaration(func) = &stmt.consequent {
-        if ptr::eq(func.as_ref(), function) {
-            return true;
-        }
-    }
-    if let Some(Statement::FunctionDeclaration(func)) = &stmt.alternate {
-        if ptr::eq(func.as_ref(), function) {
-            return true;
-        }
-    }
-    false
-}
-
 impl<'a> Binder<'a> for Function<'a> {
     fn bind(&self, builder: &mut SemanticBuilder) {
         let includes = if self.declare {
@@ -168,27 +142,19 @@ impl<'a> Binder<'a> for Function<'a> {
         };
 
         if let Some(ident) = &self.id {
-            if is_function_part_of_if_statement(self, builder) {
-                let symbol_id = builder.scoping.create_symbol(
-                    ident.span,
-                    ident.name.into(),
-                    includes,
-                    ScopeId::new(u32::MAX - 1), // Not bound to any scope.
-                    builder.current_node_id,
-                );
-                ident.symbol_id.set(Some(symbol_id));
+            let excludes = if builder.source_type.is_typescript() {
+                SymbolFlags::FunctionExcludes
+            } else if is_function_part_of_if_statement(self, builder) {
+                SymbolFlags::empty()
             } else {
-                let excludes = if builder.source_type.is_typescript() {
-                    SymbolFlags::FunctionExcludes
-                } else {
-                    // `var x; function x() {}` is valid in non-strict mode, but `TypeScript`
-                    // doesn't care about non-strict mode, so we need to exclude this,
-                    // and further check in checker.
-                    SymbolFlags::FunctionExcludes - SymbolFlags::FunctionScopedVariable
-                };
-                let symbol_id = builder.declare_symbol(ident.span, &ident.name, includes, excludes);
-                ident.symbol_id.set(Some(symbol_id));
-            }
+                // `var x; function x() {}` is valid in non-strict mode, but `TypeScript`
+                // doesn't care about non-strict mode, so we need to exclude this,
+                // and further check in checker.
+                SymbolFlags::FunctionExcludes - SymbolFlags::FunctionScopedVariable
+            };
+
+            let symbol_id = builder.declare_symbol(ident.span, &ident.name, includes, excludes);
+            ident.symbol_id.set(Some(symbol_id));
         }
 
         // Bind scope flags: GetAccessor | SetAccessor
